fix(bb): wasm-emscripten cutover — NODERAWFS link flag, host env forwarding, pthread stack, library main override

+++ barretenberg/cpp/src/barretenberg/polynomials/backing_memory.cpp
@@ -63,7 +63,10 @@ size_t parse_size_string(const std::string& size_str)
             throw_or_abort("Invalid storage size format: '" + size_str + "'. No numeric value provided");
         }
 
-        size_t value = std::stoull(str);
+        // stoull returns unsigned long long (always 64-bit); on WASM (32-bit
+        // size_t) this narrows, which -Wshorten-64-to-32 -Werror catches.
+        // Storage budgets always fit in size_t on the target.
+        size_t value = static_cast<size_t>(std::stoull(str));
         return value * multiplier;
     } catch (const std::invalid_argument&) {
         throw_or_abort("Invalid storage size format: '" + size_str + "'. Not a valid number");
